extract Stacked bar chart
